pyproject.toml
[project]
name = "polybot"
version = "0.1.0"
description = "Polymarket copy trading bot"
readme = "CLAUDE.md"
requires-python = ">=3.9"
dependencies = [
    "py-clob-client>=0.17.0",
    "requests>=2.31.0",
    "websockets>=12.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
